Philo

+++ sources/parse.c
@@ -6,7 +6,7 @@
 /*   By: msierra- <msierra-@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/01/19 13:47:37 by msierra-          #+#    #+#             */
-/*   Updated: 2022/02/26 19:27:51 by msierra-         ###   ########.fr       */
+/*   Updated: 2022/03/03 15:49:15 by msierra-         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -20,7 +20,7 @@ void	getvalues(int argc, char **argv, t_state *state)
 	state->t_die = atoi(argv[2]);
 	state->t_eat = atoi(argv[3]);
 	state->t_sleep = atoi(argv[4]);
-	state->t_init = (int)gettime();
+	state->t_init = gettime();
 	if (argc == 6)
 	{
 		if (argv[5] <= 0)
